+++ optimade/models/optimade_json.py
@@ -7,7 +7,7 @@ from typing import Any, Dict, List, Optional, Type, Union
 from pydantic import AnyHttpUrl, AnyUrl, BaseModel, EmailStr, root_validator
 
 from optimade.models import jsonapi
-from optimade.models.utils import SemanticVersion, StrictField
+from optimade.models.utils import SemanticVersion, StrictField, DataType
 
 __all__ = (
     "DataType",
@@ -24,112 +24,6 @@ __all__ = (
     "Relationship",
 )
 
-
-class DataType(Enum):
-    """Optimade Data Types
-
-    See the section "Data types" in the OPTIMADE API specification for more information.
-    """
-
-    STRING = "string"
-    INTEGER = "integer"
-    FLOAT = "float"
-    BOOLEAN = "boolean"
-    TIMESTAMP = "timestamp"
-    LIST = "list"
-    DICTIONARY = "dictionary"
-    UNKNOWN = "unknown"
-
-    @classmethod
-    def get_values(cls):
-        """Get OPTIMADE data types (enum values) as a (sorted) list"""
-        return sorted((_.value for _ in cls))
-
-    @classmethod
-    def from_python_type(cls, python_type: Union[type, str, object]):
-        """Get OPTIMADE data type from a Python type"""
-        mapping = {
-            "bool": cls.BOOLEAN,
-            "int": cls.INTEGER,
-            "float": cls.FLOAT,
-            "complex": None,
-            "generator": cls.LIST,
-            "list": cls.LIST,
-            "tuple": cls.LIST,
-            "range": cls.LIST,
-            "hash": cls.INTEGER,
-            "str": cls.STRING,
-            "bytes": cls.STRING,
-            "bytearray": None,
-            "memoryview": None,
-            "set": cls.LIST,
-            "frozenset": cls.LIST,
-            "dict": cls.DICTIONARY,
-            "dict_keys": cls.LIST,
-            "dict_values": cls.LIST,
-            "dict_items": cls.LIST,
-            "NoneType": cls.UNKNOWN,
-            "None": cls.UNKNOWN,
-            "datetime": cls.TIMESTAMP,
-            "date": cls.TIMESTAMP,
-            "time": cls.TIMESTAMP,
-            "datetime.datetime": cls.TIMESTAMP,
-            "datetime.date": cls.TIMESTAMP,
-            "datetime.time": cls.TIMESTAMP,
-        }
-
-        if isinstance(python_type, type):
-            python_type = python_type.__name__
-        elif isinstance(python_type, object):
-            if str(python_type) in mapping:
-                python_type = str(python_type)
-            else:
-                python_type = type(python_type).__name__
-
-        return mapping.get(python_type, None)
-
-    @classmethod
-    def from_json_type(cls, json_type: str):
-        """Get OPTIMADE data type from a named JSON type"""
-        mapping = {
-            "string": cls.STRING,
-            "integer": cls.INTEGER,
-            "number": cls.FLOAT,  # actually includes both integer and float
-            "object": cls.DICTIONARY,
-            "array": cls.LIST,
-            "boolean": cls.BOOLEAN,
-            "null": cls.UNKNOWN,
-            # OpenAPI "format"s:
-            "double": cls.FLOAT,
-            "float": cls.FLOAT,
-            "int32": cls.INTEGER,
-            "int64": cls.INTEGER,
-            "date": cls.TIMESTAMP,
-            "date-time": cls.TIMESTAMP,
-            "password": cls.STRING,
-            "byte": cls.STRING,
-            "binary": cls.STRING,
-            # Non-OpenAPI "format"s, but may still be used by pydantic/FastAPI
-            "email": cls.STRING,
-            "uuid": cls.STRING,
-            "uri": cls.STRING,
-            "hostname": cls.STRING,
-            "ipv4": cls.STRING,
-            "ipv6": cls.STRING,
-        }
-
-        return mapping.get(json_type, None)
-
-
-class AllowedJSONSchemaDataType(Enum):
-    """The allowed values for `type` in the Property Definition restricted JSON Schema syntax."""
-
-    BOOLEAN = "boolean"
-    OBJECT = "object"
-    ARRAY = "array"
-    NUMBER = "number"
-    STRING = "string"
-    INTEGER = "integer"
 
 
 class OptimadeError(jsonapi.Error):
